+++ praktikum/burger.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from src.bun import Bun
-from src.ingredient import Ingredient
+from praktikum.bun import Bun
+from praktikum.ingredient import Ingredient
 
 
 class Burger:
